requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black
asgiref==3.3.1            # via django
black==20.8b1             # via -r requirements.in
certifi==2020.11.8        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools
django-environ==0.4.5     # via -r requirements.in
django==3.1.4             # via -r requirements.in
idna==2.10                # via requests
mypy-extensions==0.4.3    # via black
pathspec==0.8.1           # via black
pillow==8.0.1             # via -r requirements.in
pip-tools==5.4.0          # via -r requirements.in
python-dotenv==0.15.0     # via -r requirements.in
pytz==2020.4              # via django
regex==2020.11.13         # via black
requests==2.25.0          # via -r requirements.in
six==1.15.0               # via pip-tools
sqlparse==0.4.1           # via django
toml==0.10.2              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.26.2           # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
